Use a webhook to automatically update the static site content
Prompt:
Move the ghost static site generator to a container defined in the main docker-compose.yml file, which is always running, and to generate the static sites, instead of running a new image, use docker exec to run the same command within that docker container.
Add a simple webhook receiver, running in a docker container defined in the main docker-compose.yml site, that will receive a `site.changed` webhook from any of the ghost sites whenever any changes are made to them, and trigger the static site generator for the appropriate site.
After a new site has been created, tell the user how to set up the webhook on that site.
If possible, make this webhook receiver only accessible by local docker containers rather than publicly.

(And then told it to deduplicate the README)